Doc to knowledge graph generator/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
rdflib
google-generativeai
PyPDF2
python-docx
